+++ src/Algebra/Graph/IO/Datasets/LINQS/Citeseer.hs
@@ -7,6 +7,8 @@
 -- Qing Lu, and Lise Getoor. "Link-based classification." ICML, 2003.
 --
 -- https://linqs.soe.ucsc.edu/data
+--
+-- The dataset consists of 3312 scientific publications classified into one of six classes. The citation network consists of 4732 links. Each publication in the dataset is described by a 0\/1-valued word vector indicating the absence\/presence of the corresponding word from the dictionary. The dictionary consists of 3703 unique words.
 module Algebra.Graph.IO.Datasets.LINQS.Citeseer (
   -- * 1. Download the dataset
   stash
@@ -74,16 +76,16 @@ stash :: FilePath -- ^ directory where the data files will be saved
 stash fp = DL.stash fp "http://www.cs.umd.edu/~sen/lbc-proj/data/citeseer.tgz" 3703 docClassP
 
 -- | See `DL.sourceGraphEdges`
-sourceCiteseerGraphEdges :: (MonadResource m, MonadThrow m) =>
+sourceCiteseerGraphEdges :: (MonadResource m, MonadThrow m, Binary ix) =>
                       FilePath -- ^ directory of data files
-                   -> M.Map String (Seq Int16, CiteSeerDoc) -- ^ 'content' data
-                   -> ConduitT i (Maybe (G.Graph (DL.ContentRow CiteSeerDoc))) m ()
+                   -> M.Map String (ix, Seq Int16, CiteSeerDoc) -- ^ 'content' data
+                   -> ConduitT i (Maybe (G.Graph (DL.ContentRow ix CiteSeerDoc))) m ()
 sourceCiteseerGraphEdges = DL.sourceGraphEdges
 
 -- | See `DL.loadGraph`
-loadCiteseerGraph :: 
+loadCiteseerGraph :: (Binary ix) => 
                      FilePath -- ^ directory where the data files were saved
-                  -> IO (G.Graph (DL.ContentRow CiteSeerDoc))
+                  -> IO (G.Graph (DL.ContentRow ix CiteSeerDoc))
 loadCiteseerGraph = DL.loadGraph
 
 -- | document classes of the Citeseer dataset